Enabled to be detected multiple words.
In the conventional format of `rules`, it is necessary to increase `pattern` for each words to be detected.
Therefore, when many words to be detected, the number of lines in the `rule` increased, and  it make to be difficult to read.
To solve this problem, we made it possible to detect many words in a `rule`.